A little bit of deBugging

+++ src/AlphabetValues.java
@@ -1,7 +1,7 @@
 public class AlphabetValues {
 
     private static char[] alphabet = {'A','B','C','D','E','F','G','H','I','J','K','L','M','N','O','P','Q','R','S','T','U','V','W','X','Y','Z'};
-    private static int[] abcValue = {10, 11, 12, 13, 14,15,16,17,18,19,20,21,22,23,24,25,26,27,28,29,30,31,32,34,35};
+    private static int[] abcValue = {10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21, 22, 23, 24, 25, 26, 27, 28, 29, 30, 31, 32, 34, 35};
 
     public static int getLetterValue(char letter) {
         if (letter == 'A') {
@@ -25,7 +25,7 @@ public class AlphabetValues {
             return abcValue[i];
         }
     }
-    public static String  getLetterFromValue(long value) {
+    public static String  getLetterFromValue(int value) {
         if (value == 10) {
             return "A";
         } else {
